added Cygnet Skywalker v1 PCB Project
Cygnet Skywalker v1 is the first iteration of a Payload Computer PCB board with a RADFET (Radiation Field Effect Transistor) and Space-grade GPS transceiver along with state of health telemetry sensors. It communicates with the main flight computer board and transmits data for downlinking and receives power from a battery board.